move dfschema and column
